Complete Storage Implementation with manual testing
