+++ data/duke.txt
@@ -1,2 +1,3 @@
 T | 0 | read book
-E | 0 | NUS GreyHats | 19/09/2023 1800 | 21/09/2023 1800
+E | 1 | NUS GreyHats | 19/09/2023 1800 | 21/09/2023 1800
+D | 0 | submit lab | 19/08/2023 1900
